Update favicon
The portal should use the dansdata favicon.

This commit adds the favicon to the default document rendered by
nextjs. This reduces the amount of copy-paste when new pages are
added in the future.

The html `lang` attribute is ignored in this commit as I have not
yet decided on a render strategy. Hopefully nextjs will manage
this for us but that is as of yet not confirmed.

See also https://nextjs.org/docs/advanced-features/custom-document