Configure StorybookJS
[Storybook](https://storybook.js.org/) allows us to document individual
components and makes testing easier.

This commit adds and configures storybook. This process turned out to be
a little more tedious than expected due to some minor issues with
webpack versioning (no idea why - explicitly locking webpack@5 solved
this) as well as SSR and and next-i18next (see
https://github.com/storybookjs/storybook/issues/4082#issuecomment-818032557
)

The eslint configuration file was converted to js in order to support
Storybook's automated eslint migration.

I found this SO question useful during the configuration:
https://stackoverflow.com/questions/69810773/next-i18next-translation-not-working-in-storybook-console-logs-missingkey