Set up typography and add typography section to storybook
Following the guide at
https://m3.material.io/styles/typography/type-scale-tokens
I added design tokens for typography. This is probably overkill but I
decided to do it as a fun excersize.

In this commit I have selected
[`Norican`](https://fonts.google.com/specimen/Norican) as the brand font
and [`Roboto Flex`](https://fonts.google.com/specimen/Roboto+Flex) as
the "plain" font.

Initially, I was going to go with
[`Qwigley`](https://fonts.google.com/specimen/Qwigley) as the brand font
but I found the font a bit hard to read at smaller sizes.

Next.JS apparently has built in web font optimization:
https://nextjs.org/docs/basic-features/font-optimization . For this
reason, I decided not to include the `<link rel="preconnect" .../>` tags
otherwise suggested by google. I'm not sure if this is better or not
but the example at nextjs.org did not have them.