Improve button styling and prepare for more components
After finally getting back to this project after a few weeks,
it seems crazy not to provide react components to wrap the
component classes (`.btn` etc.).

This commit rewrites some of the `button.css` (now `button/index.css`)
styling based on the assumption that child nodes will always be full
elements. The commit also documents this assumption (and a few other
things) and adds a helper react component to make adding buttons easy.

Finally, the directory structure is changed slightly to support more
components in an organized manner.