Expand clickable area of text fields
The clickable area of the input/textarea is considerably smaller than
the text fields themselves. This is hard to adjust for without using
javascript.

This commit attempts to reduce the issue by using paddings instead of
margins on the input/textarea elements, as clicking the padding is still
considered a click on the element (whereas a click on a margin is not).