Use .env.development for supabase variables
The local supabase setup will be the same, including anon key, for every
developer. To make getting started easier, it is better to keep these
variables in the repo.

Note that the ANON_KEY for local development has a constant value of
"super-secret-jwt-token-with-at-least-32-characters-long" as an
undocumented "feature" (as described in this comment:
https://github.com/supabase/supabase-js/issues/25#issuecomment-1019935888
)