Add LanguageButton for language switching
We'll want to keep the logic for changing language centralized. A
specialized button implementation for this purpose should be perfect.

NextJS requires that complex (i.e. custom) components that are children
to `<Link>` tags can accept a ref. `<Button>` is therefore updated to
support this.
https://nextjs.org/docs/api-reference/next/link#if-the-child-is-a-functional-component

The change from `function Button` to `const Button =` was to appease
storybook.