Set up tbls to auto-generate database documentation
While Supabase Studio provides a good overview of individual tables, it
will probably be beneficial for contributors, old and new, to have
access to a visual representation as well.

This commit sets up [tbls](https://github.com/k1LoW/tbls) to generate
database diagrams and github-friendly markdown files. Prettier was
suppressed for the `docs/database` folder to prevent interference with
the auto-generated files.