Add support for API tokens
While supabase supports authentication out of the box, it does not
support the generation of API key. After searching for quite a bit,
I opened a feature request with Supabase
(https://github.com/supabase/supabase/issues/7186) and made a deep dive
into Supabase, PostgREST, etc. to come up with the solution in this
commit.

The migration is written in an article-like fashion to try and see if I
liked such writing. The finished code was also released on
https://gist.github.com/FelixZY/0aef530690458b381b8100afa19202c8.

It should be noted that no environments beyond "local" has been set up
yet, since I have not yet set up any remote supabase environments for
this project. Placeholders are therefore left in place for now.