Convert translation files to yaml
It will be necessary to have longer blocks of text in the translation
files, sometimes including markdown. This is really hard to read and
write in JSON whereas in YAML we can use multiline strings.

To make things easier, this commit switches from json to yaml
translation files.

Note that `npm run dev` causes a warning:
```
warn  - Invalid next.config.js options detected:
  - The root value has an unexpected property, localePath [...]
  - The root value has an unexpected property, localeExtension [...]
```
However, these options are valid according to
https://github.com/i18next/next-i18next and vercel does indeed have some
problems with the translation files if `localePath` is not specified.