Set up github actions for database migrations
Based on the official documentation here:
https://supabase.com/docs/guides/cli/cicd-workflow

This commit introduces a slight naming issue with the terms "preview"
and "staging". The origin of this conflict is that vercel does not
separate between "preview" and "staging" and originally, I was thinking
there did not need to be a difference. However, after doing some more
work, it turns out that SSO will be broken without a fixed return url
(which requires a fixed branch in vercel). Also, after thinking things
through a bit more, it does not make sense to allow just any PR to apply
database changes. Therefore the term "staging" will be used to refer to
a preview deployment originating from the `staging` branch. This preview
deployment will be allowed to apply database migrations and given a
fixed url in order to e.g. enable SSO.