Add circular progress indicator
As specified on
https://m3.material.io/components/progress-indicators/specs

The plan was to implement the progress indicator using CSS. However, the
animation turned out to be a bit too complex without involving
`@property` variables which are not supported e.g. by firefox. For this
reason, I chose to fall back on svg with `<animate>` tags. This seems to
produce a smoother result than my CSS version but does not keep the
stroke at 4dp when resizing the indicator.

Note that the indicator also includes a 4dp padding. This is due to
[this](
https://lh3.googleusercontent.com/pMrGgRn-O92mCRx34N4rQ3zuTjZMYYBezpOFflBYI0nm5n9zc2Y4sHqa9jcWMGRt-Ot7KvI5eskA0K5yaSJrx6-MxrboQtd4xteLmQf1dyc=s0
) image from the specification including some additional padding which
seems to be undocumented. 4dp is a guess and may even be too small.

The animation implemented in this commit relies on one linear and one
easing function working together. The result is based on eyeballing the
example from the specification and is not necessarily 100% correct.