feat: enable Facebook, GitHub and Google SSO with Supabase
It should be easy to use Danstavlan. Therefore, users should be able to
use their existing accounts and sign in via SSO.

This commit enables SSO for the local Supabase environment (a frontend
still needs to be implemented). In conjunction with this commit, I have
also configured environment variables in vercel to make SSO available in
production as well.

SSO in preview deployments are currently not solved. The issue here is
that the preview deployment url is generated and cannot be be provided
to Facebook, Google or GitHub beforehand.