feat: add pages for legal information
E.g. Facebook requires a privacy policy to implement their SSO flow.
While this policy is not yet written, we will need one. We will also
need a terms of service document to prevent abuse.

I'm not a lawyer and writing legal texts in one language is hard enough.
Therefore, the legal texts will only be available in swedish.

The prose styling is also updated as I've now been able to see it in
action with long texts.