Add `associations` tables
Dance events are quite dynamic. One event may have 2 bands, one may
have 3. In some cases, the band is also the organizer, an instructor may
take photos, etc. For this reason, I have opted for a very dynamic
approach where relationships between profiles and event_slots are based
on "association tags". This allows any profile to have any role.

There will probably need to be some limits to this - a venue will need
to provide a location for example - but that is a problem for later.