feat: add main pages for navigation when signed in
The main pages for a signed-in user is expected to be:
* Events - Where users can manage upcoming dances and related events
* Profiles - Where users can manage e.g. band profiles
* API - Where users can generate API keys and manage their API usage
* Account - Where users can manage their account settings

This commit adds these pages in preparation for setting up a navigation
component.

Note that the new page-specific translations are located in
`${page}/page.yaml`. This is in preparation for a future migration to
Next 13's `app` directory where the page will be located in
`app/secure/${path}/page.tsx`. This is also why each introduced page is
placed in separate directories like `secure/api/index.tsx` instead of as
`secure/api.tsx`.