feat: implement navigation layout for logged in state
Set up:
* Navigation drawer
  For desktop layouts
  https://m3.material.io/components/navigation-drawer
* Navigation rail
  For tablet layouts
  https://m3.material.io/components/navigation-rail
* Navigation bar
  For mobile layouts
  https://m3.material.io/components/navigation-bar

as primary navigation for logged in users.

Previously, I've implemented components css-first. However, this has
caused a lot of unnecessary overhead and slowdown in a phase where I
want to move more quickly. This commit therefore implements the
components directly in react using tailwind classes for styling.

This commit also adds a top app bar for mobile layouts in order to fit
the dark mode toggle and logout functionality. However, this top app bar
is purpose-built to provide mobile users with the same functionality as
tablet and desktop users and may not be dynamic enough for our future
needs. Time will tell (good enough for now).

Finally, the new `NavigationLayout` requires knowledge of the user's
device size to select the correct component. However, this information
is not available during SSR and the component therefore falls back on
not rendering any layout. This causes a small jump during page load
which is undesirable. For now I've deemed it acceptable but in the
future we may want to migrate to a more css-heavy solution to avoid
this. I'm also thinking Next 13 with client components may be able to
work around this issue.