feat: set color-scheme to reduce blank flash while loading flutter
Flutter renders its content in a canvas. While the canvas is being
prepared, a blank page is displayed. This page is typically white,
causing a short flash for dark-mode devices.

By setting `color-scheme` on the root element to indicate that both
light and dark mode is supported, this color flash can be minimized as
the user's browser can adapt to the system settings before we can. This
does not take user overrides within the portal into account but should
be good enough.