chore(devcontainer): stop installing golang in devcontainer
Go was originally included since this project uses some libraries based
on go. The idea was that including go would make interaction with the
source code of such libraries easier if needed. However, such
interaction is very rarely required and the inclusion of golang mostly
increases the devcontainer build time.