feat: add language and theme mode settings
After playing around with some different ways of managing settings, I
think this is the way to go.

The project will use
[provider](https://pub.dev/packages/provider) for dependency injection.
This allows individual classes to be injected deep in the hierarchy and
for widgets to update only if the class they are watching has been
updated (via `ChangeNotifier`). This also means we want to avoid
collecting multiple settings in one file (`a la `SettingsController`)
since this could trigger unnecessary updates when a single setting is
changed.

This commit configures a `Setting` base class, to help with common
persistance operations, as well as two settings: one for language and
one for theme mode. Note that these cannot yet be toggled in the ui.