feat: show splash screen to enable code splitting
Time to first paint is important on the web. For this reason, we want
our initial download to be as small as possible, meaning we need to
employ code splitting.

As I understand it, flutter supports code splitting - but only if
packages are imported using the `deferred as` syntax. This commit
therefore sets up a minimal splash screen, allowing the rest of the app
to be loaded while the splash screen is showing.

Note that the `SplashScreenModule` is nearly identical to `AppModule` at
this time. However, as more dependencies are introduced, the differences
will grow and the `SplashScreenModule` can remain minimal.