feat: allow page title to be controlled from anywhere
We'll want to be able to update the page title from different pages in
the application. This means the page title must be made available to the
top level `MaterialApp`. This calls for injection. However, `String` is
not a suitable type since there could then not be any other string-value
injections. `Setting` has a suitable API but we don't want the page
title persisted to disk.

This commit introduces a solution in the form of `State` - a base class
for wrapping mutable values in an injectable container. The basic API of
`Setting` is moved to `State` and `Setting` is made a specialization of
`State`, for persistent application settings.

This commit also introduces a new PageTitleState, making it possible to
modify the page title from anywhere, as desired.