fix: prevent accidental wrong color in logo title/tagline
Some widgets modify the theme text styles from the "root" styles. This
can cause "accidental" overrides of the text in `DansdataLogo`.

Currently, it seems impossible to reset the `TextStyle.color` value to
`null` once it has been set. This commit therefore solves the problem by
introducing `Colors.transparent` as a "flag" color, to indicate that no
override was intended by the caller.