build(flutter): bump flutter to 3.7.0-1.4.pre
Flutter's support for MD3 is not yet complete, but a lot has happened in
the last quarter (flutter releases are ~quarterly). Lately, I've found
myself looking at a lot of "merged" PRs that are available in the beta
channel but not yet in stable.

Given the time since the last stable and the upcoming Flutter Forward
event on 2023-01-25, I'm expecting Flutter 3.7 stable to drop any day
now. The latest 3.7 beta should therefore be ok to use early in order
not to block development efforts.

In the related upgrade to dart 2.19, the analyzer warned that the
`invalid_use_of_protected_member` analyzer rule is not a valid entry for
`linter: rules:`. I therefore removed this rule. See
https://github.com/dart-lang/sdk/issues/51090 for more details.

See also:
* https://github.com/flutter/flutter/pull/107382
* https://github.com/flutter/flutter/issues/91605