fix(PortalAppBar): disable logo wrapping
The logo should never overlap with the action buttons. However, if this
would happen, it is better that we get runtime warnings in debug builds
(flutter will raise warnings about overflowing content e.g. in widget
tests). The nowrap version of the logo also looks better in case of an
overflow.