feat: add `ConditionalWrapper` widget
This widget makes it easier to wrap components _under certain
circumstances_. For example, we want all text in the application to be
selectable on web (as this is expected platform behavior) but not in
native apps. By using `ConditionalWrapper` we can avoid code duplication
and make our intent more easily understood.