refactor: rethink ViewModel initialization
I've been unsure about whether to use `addPostFrameCallback` as well as
where to handle the fact that you cannot change signals that affect a
parent widget during `initState` and `didChangeDependencies` as that
causes a "setState() or markNeedsBuild() called during build" error to
be raised.

After some [feedback from @jinyus](
https://github.com/rodydavis/signals.dart/discussions/142#discussioncomment-8362517
), it seems I was still pretty close with my initial implementation.
However, based on the feedback I was still able to refine my
understanding and come up with the following "ruleset":
* Mutating the inner state of the ViewModel in `initState` or
  `didChangeDependencies` is fine.
* Mutating external state cannot be done in `initState` or
  `didChangeDependencies`. Instead, `addPostFrameCallback` is required
  to "bail out" of the build first.
  * Technically, it is fine to mutate external state that does not cause
    parent widgets to change but it is safer and more consistent for the
    ViewModel to assume that all external state is "dangerous".
* As the aforementioned issues are tied to how widgets are built, the
  widget (or its state) should be responsible for the call to
  `addPostFrameCallback` - not the ViewModel.

With this commit, I'm splitting the old `init` function into two,
`initLocalState` and `initGlobalState`. This allows ViewModels to
configure locals pre-frame and globals post-frame. Hopefully the naming
will also make it easier to understand where to do what.